requirements.txt
requests
allure-pytest
pytest~=7.3.1
curlify
jsonschema
Pillow~=10.0.1